requirements.txt
django
djangorestframework
django-money